# data-ref/src/main/resources/META-INF/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API Baseline
  description: CRUD Operation on repositories data
  version: 0.7.0
servers:
  - url: http://data-ref.provoly.com
paths:
...
  /abac/rules:
    get:
      tags:
        - Abac Controller
      summary: Retrieve all ABAC rules
      description: |
        The ABAC (Attribute-Based Access Control) are rules and predicates.
        They allow us to provide access to users based on who they are rather than what they do.
      operationId: com.provoly.ref.abac.AbacController_getRules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AbacRuleDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - data_access_read
    post:
      tags:
        - Abac Controller
      summary: Save a new ABAC rule
      operationId: com.provoly.ref.abac.AbacController_addRule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbacRuleDto'
            examples:
              attributeAdd:
                summary: Adding an ABAC rule of type attribute
                value: |
                  {
                    "id": "077e4ac8-7499-4489-bd5f-ac37d2ca226d",
                    "type": "ATTRIBUTE",
                    "name": "Attribute example",
                    "predicate": "077e4ac8-7499-4489-bd5f-ac37d2ca226e",
                    "condition": {
                      "type": "ATTRIBUTE",
                      "attribute": "f897dc76-9c4a-4b7a-9797-fc577c0438e9",
                      "operator": "EQUALS",
                      "value": 1234567890
                    }
                  }
              composedAdd:
                summary: Adding a composed ABAC rule
                value: |
                  {
                    "id": "7ffcea67-2335-429f-888a-59121f522144",
                    "name": "MyRule",
                    "type": "ATTRIBUTE",
                    "predicate": "7ffcea67-2335-429f-888a-59121f522143",
                    "condition": {
                      "type": "OR",
                      "composed": [
                        {
                          "type": "AND",
                          "composed": [
                            {
                              "type": "ATTRIBUTE",
                              "oClass": "411a19f5-9d00-4028-b4b6-4b04101f6316",
                              "attribute": "9d82213c-db91-42db-8675-5f2e70ae2f05",
                              "value": 202000010265
                            },
                            {
                              "type": "ATTRIBUTE",
                              "oClass": "411a19f5-9d00-4028-b4b6-4b04101f6316",
                              "attribute": "dbf3917f-1af8-495e-94af-67d8315a7772",
                              "value": 3
                            }
                          ]
                        },
                        {
                          "type": "ATTRIBUTE",
                          "oClass": "411a19f5-9d00-4028-b4b6-4b04101f6316",
                          "attribute": "9d82213c-db91-42db-8675-5f2e70ae2f05",
                          "value": 202000010265
                        },
                        {
                          "type": "ATTRIBUTE",
                          "oClass": "411a19f5-9d00-4028-b4b6-4b04101f6316",
                          "attribute": "dbf3917f-1af8-495e-94af-67d8315a7772",
                          "value": 3
                        }
                      ]
                    }
                  }
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          content:
            application/json:
              examples:
                differentOClass:
                  description: Different OClass in AND/OR condition
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "Attributes are not on same OClass"
                    }
                differentTypes:
                  description: If the ABAC rule type is different from its condition type
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "RuleType ATTRIBUTE does not allowed Metadata condition"
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "501":
          description: Unsupported Operation. In the case you try to add an ABAC rule with a condition we do not support.
          content:
            application/json:
              examples:
                unknownType:
                  value: |
                    {
                      "application": "data-ref",
                      "code": "NOT_IMPLEMENTED",
                      "message": "Unknown condition type {{condition.type}}"
                    }
      security:
        - oidc:
            - data_access_write
  /abac/rules/class/{oClassId}:
    get:
      tags:
        - Abac Controller
      summary: Retrieve all ABAC rules for a class based on its identifier
      operationId: com.provoly.ref.abac.AbacController_getRulesFor
      parameters:
        - name: oClassId
          in: path
          description: Class UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: includeInactive
          in: query
          description: Boolean to retrieve inactive rules.
          schema:
            type: boolean
          example: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AbacRuleDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - data_access_read
            - search
            - datasource_read
  /abac/rules/id/{ruleId}:
    get:
      tags:
        - Abac Controller
      summary: Retrieve an ABAC rule based on its identifier
      operationId: com.provoly.ref.abac.AbacController_getRule
      parameters:
        - name: ruleId
          in: path
          description: ABAC rule UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbacRuleDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - data_access_read
  /abac/rules/{ruleId}:
    delete:
      tags:
        - Abac Controller
      summary: Delete an ABAC rule based on its identifier
      operationId: com.provoly.ref.abac.AbacController_deleteRule
      parameters:
        - name: ruleId
          in: path
          description: UUID of the rule
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - oidc:
            - data_access_write
...
components:
  responses:
    notFoundException:
      description: Entity not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: 404
        message:
          type: string
          example: "Entity : ac2c1a7f-0694-4f03-bb00-d349adeecd3d inexistant."
      required:
        - code
        - message
    AbacRuleDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "my rule"
        description:
          type: string
          example: "description of my rule"
        active:
          type: boolean
          example: true
          description: If abac rule is not active, then it will not be taken into account for restriction processes.
        predicate:
          $ref: '#/components/schemas/UUID'
        type:
          $ref: '#/components/schemas/AbacRuleType'
        condition:
          $ref: '#/components/schemas/ConditionDto'
    AbacRuleType:
      enum:
        - METADATA
        - ATTRIBUTE
      type: string
...
    ConditionDto:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ConditionType'
    ConditionType:
      enum:
        - "TRUE"
        - OR
        - AND
        - ATTRIBUTE
        - METADATA
      type: string
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
  securitySchemes:
    oidc:
      type: openIdConnect
      description: Authentication
      openIdConnectUrl: https://auth/realms/provoly/.well-known/openid-configuration